WIP: Add spherical harmonic transformations
